+++ src/components/EditTestsSelector/EditTestsFilter/EditTestsFilter.jsx
@@ -4,7 +4,7 @@ import InputLabel from '@material-ui/core/InputLabel';
 import FormControl from '@material-ui/core/FormControl';
 import TextField from '@material-ui/core/TextField';
 import Select from '@material-ui/core/Select';
-import './EditTestsInfo.scss';
+import './EditTestsFilter.scss';
 import MenuItem from '@material-ui/core/MenuItem';
 import { useFormik } from 'formik';
 import PropTypes from 'prop-types';
@@ -14,7 +14,7 @@ const validation = Yup.object({
   questId: Yup.number()
 });
 
-const EditTestsInfo = (props) => {
+export const EditTestsFilter = (props) => {
 
   const testLevels = ['A1', 'A2', 'B1', 'B2', 'C1', 'C2'];
   const testModules = ['Grammar', 'Listening', 'Essay', 'Speaking'];
@@ -66,15 +66,13 @@ const EditTestsInfo = (props) => {
           Search
         </Button>
       </form>
-        {formik.errors.questId ? <div>In field 'Question Id' must be a number</div> : null}
+      {formik.errors.questId ? <div>In field 'Question Id' must be a number</div> : null}
     </>
   );
 };
 
-EditTestsInfo.propTypes = {
+EditTestsFilter.propTypes = {
   setLevel: PropTypes.func,
   setModule: PropTypes.func,
   setQuestId: PropTypes.func
 };
-
-export default EditTestsInfo;
