+++ src/components/EditTestsSelector/EditTestsSelector.jsx
@@ -1,7 +1,7 @@
 import React, { useState } from 'react';
-import EditTestsInfo from './EditTestsInfo/EditTestsInfo';
+import { EditTestsFilter } from './EditTestsFilter/EditTestsFilter';
 import imageSrc from '../../assets/images/goose.svg';
-import EditTestsData from '../EditTestsSelector/EditTestsData/EditTestsData';
+import { EditTestsTable } from './EditTestsTable/EditTestsTable';
 import './EditTestsSelector.scss';
 
 const EditTestsSelector = () => {
@@ -12,18 +12,15 @@ const EditTestsSelector = () => {
 
   return (
     <div className='edit-tests-selector'>
-      <EditTestsInfo setLevel={setLevel} setModule={setModule} setQuestId={setQuestId} />
+      <EditTestsFilter setLevel={setLevel} setModule={setModule} setQuestId={setQuestId} />
       {
         level || module || questId
-          ? <EditTestsData level={level} module={module} questId={questId} />
+          ? <EditTestsTable level={level} module={module} questId={questId} />
           : <div className='edit-tests-placeholder'>
             <div className='edit-tests-selector-text'>Select the option at the top to see the questions.</div>
             <img title='goose-img' alt='goose-img' src={imageSrc} className='picture' />
           </div>
       }
-
-
-
     </div>
   );
 };
