+++ src/components/EditTestsSelector/EditTestsTable/EditTestsTable.scss
