+++ src/store/store.js
@@ -4,11 +4,13 @@ import { combineReducers } from 'redux';
 import { jwtReducer } from './reducers/jwtReducers';
 import { authReducer } from './reducers/authReducers';
 import { profileReducer } from './reducers/profileReducers';
+import { coachReducer } from './reducers/coachReducer';
 
 const rootReducers = combineReducers({
   profile: profileReducer,
   auth: authReducer,
   jwt: jwtReducer,
+  coach: coachReducer
 });
 const composeEnhancers = window.__REDUX_DEVTOOLS_EXTENSION_COMPOSE__ || compose;
 const store = createStore(rootReducers, composeEnhancers(applyMiddleware(thunk)));
